prp-agent/pyproject.toml
[project]
name = "prp-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "openai>=1.98.0",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.4.11",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sentry-sdk[fastapi]>=2.34.1",
    "uvicorn>=0.35.0",
]
